feat: Add "Add Event" button
Introduces a new "Add Event" button on the Events page. This button will allow users to easily navigate to a page for submitting new events to the Baguio Startup Network, enhancing community participation and content generation.

The associated iCal download functionality has been removed as it was not fully implemented.